+++ autoprotocol/builders.py
@@ -23,7 +23,7 @@ See Also
 Instruction
     Instructions corresponding to each of the builders
 """
-from collections import defaultdict
+from collections import OrderedDict, defaultdict
 from collections.abc import Iterable  # pylint: disable=no-name-in-module
 from functools import reduce
 from numbers import Number
@@ -1531,9 +1531,15 @@ class LiquidHandleBuilders(InstructionBuilders):
 
     @staticmethod
     def device_mode_params(
+        device: str,
         model: Optional[str] = None,
         chip_material: Optional[str] = None,
         nozzle: Optional[str] = None,
+        diaphragm: Optional[int] = None,
+        nozzle_size: Optional[Union[Unit, str]] = None,
+        tubing: Optional[str] = None,
+        z_drop: Optional[Union[Unit, str]] = None,
+        viscosity: Optional[str] = None,
     ):
         """Helper for building device level mode_params
 
@@ -1542,12 +1548,25 @@ class LiquidHandleBuilders(InstructionBuilders):
 
         Parameters
         ----------
-        model : string, optional
-            Tempest chip model.
-        chip_material : string, optional
+        device: str
+            x_tempest_chip or x_mantis
+        model: Optional[str]
+            Tempest chip or mantis model.
+        chip_material: Optional[str]
             Material that the tempest chip is made of.
-        nozzle : string, optional
-            Type of chip nozzle.
+        nozzle: Optional[str]
+            Type of chip nozzle for tempest.
+        diaphragm: Optional[Integer]
+            any integer between 0 and 100, inclusive
+        nozzle_size: Optional[Unit[Length]] or Optional[str]
+            one of "0.1:mm", "0.2:mm", "0.5:mm"
+        tubing: Optional[String]
+            one of "LV", "HV", "P200", "P1000"
+            TODO: support that source is pipette-tip type
+        z_drop: Optional[Unit[Length]] or Optional[str]
+            within the range: 0:mm - 100:mm, inclusive
+        viscosity: Optional[String]
+            one of "1", "2-5", "6-10", "11-20", "21-25"
 
         Returns
         -------
@@ -1557,38 +1576,119 @@ class LiquidHandleBuilders(InstructionBuilders):
         Raises
         ------
         ValueError
-            If model is not None or "high_volume"
-        ValueError
-            If nozzle is not None or "standard"
-        ValueError
-            If chip_material is not None or in the allowable list of tempest chip materials
+            If device is not of the following: [x_tempest_chip, x_mantis]
         """
-        # tempest chip specification
-        if (model != "high_volume") and (model is not None):
-            raise ValueError(f"Tempest chip model is {model}. It must be: high_volume.")
-        if (nozzle != "standard") and (nozzle is not None):
-            raise ValueError(f"Tempest nozzle is {nozzle}. It must be: standard.")
-        if (chip_material not in ["silicone", "pfe"]) and (chip_material is not None):
-            raise ValueError(
-                f"Tempest chip material is {chip_material}. It must be either: silicone or pfe."
-            )
-
-        device_mode_params = {}
-        if any([model, chip_material, nozzle]):
-            if model is None:
-                model = "high_volume"
-            if chip_material is None:
-                chip_material = "pfe"
-            if nozzle is None:
-                nozzle = "standard"
-            x_tempest_chip = {
+        params: OrderedDict = OrderedDict(
+            {
                 "model": model,
                 "material": chip_material,
                 "nozzle": nozzle,
+                "diaphragm": diaphragm,
+                "nozzle_size": nozzle_size,
+                "tubing": tubing,
+                "z_drop": z_drop,
+                "viscosity": viscosity,
             }
-            device_mode_params["x_tempest_chip"] = x_tempest_chip
+        )
+        if device not in ["x_mantis", "x_tempest_chip"]:
+            raise ValueError(
+                f"Device is {device}. It must be: [x_tempest_chip, x_mantis]"
+            )
+        if any(params.values()):
+            LiquidHandleBuilders.validate_device_params(device, params)
 
-        return device_mode_params
+        device_dict: dict = {}
+        for key, value in params.items():
+            if value or isinstance(value, int):
+                if key == "nozzle_size" or key == "z_drop":
+                    device_dict.update({key: Unit(value)})
+                else:
+                    device_dict.update({key: value})
+
+        return {device: device_dict}
+
+    @staticmethod
+    def validate_device_params(device: str, device_dict: dict) -> None:
+        """
+        Helper validation function to validate device liquid handling params
+        Parameters
+        ----------
+        device: str
+            either x_mantis or x_tempest_chip
+        device_dict: dict
+            Dictionary of all device params, as seen in function device_mode_params
+
+        Raises
+        ------
+        ValueError
+            If input device is not either x_mantis or x_tempest_chip
+        ValueError
+            If input device_dict values are not of the accepted
+            params for the input device
+        KeyError
+            If input device_dict key is not in accepted_params keys
+        """
+        if device == "x_mantis":
+            # If device is mantis, set accepted params
+            accepted_params: dict = {
+                "model": ["high_volume", "low_volume"],
+                "diaphragm": [0, 100],
+                "nozzle_size": [Unit("0.1:mm"), Unit("0.2:mm"), Unit("0.5:mm")],
+                "tubing": ["LV", "HV", "P200", "P1000"],
+                "z_drop": [Unit("0.0:mm"), Unit("100.0:mm")],
+                "viscosity": ["1", "2-5", "6-10", "11-20", "21-25"],
+            }
+        elif device == "x_tempest_chip":
+            # Otherwise, default to tempest accepted params
+            accepted_params: dict = {
+                "model": ["high_volume"],
+                "nozzle": ["standard"],
+                "material": ["pfe", "silicone"],
+            }
+        else:
+            raise ValueError(
+                f"Device is {device}. It must be: [x_tempest_chip, x_mantis]"
+            )
+        error_values: dict = {}
+        error_keys: dict = {}
+        # Validate params with accepted params dict
+        for key, value in device_dict.items():
+            if value:
+                if key not in accepted_params:
+                    error_keys.update({key: value})
+                elif key == "diaphragm":
+                    accepted_range: List = list(accepted_params[key])
+                    if value < accepted_range[0] or value > accepted_range[1]:
+                        error_values.update({key: value})
+                elif key == "z_drop":
+                    accepted_range: List = list(accepted_params[key])
+                    if isinstance(value, str):
+                        compare_val = Unit(value)
+                    else:
+                        compare_val = value
+                    if (
+                        compare_val < accepted_range[0]
+                        or compare_val > accepted_range[1]
+                    ):
+                        error_values.update({key: value})
+                elif key == "nozzle_size":
+                    if (
+                        isinstance(value, str)
+                        and Unit(value) not in accepted_params[key]
+                    ):
+                        error_values.update({key: value})
+                else:
+                    if value not in accepted_params[key]:
+                        error_values.update({key: value})
+
+        if error_keys:
+            raise KeyError(
+                f"Incorrect key values: {error_keys}. Accepted keys are: {accepted_params.keys()}"
+            )
+        if error_values:
+            raise ValueError(
+                f"Incorrect params: {error_values}. It must be {accepted_params}"
+            )
 
     @staticmethod
     def move_rate(